call-center-api/requirements.txt
aiohttp==3.9.1
aiosignal==1.4.0
alembic==1.16.5
amqp==5.3.1
annotated-types==0.7.0
anyio==3.7.1
asgiref==3.5.0
async-timeout==4.0.3
asyncpg==0.30.0
attrs==25.3.0
babel==2.17.0
bcc==0.18.0
bcrypt==4.3.0
billiard==4.2.1
black==23.11.0
bleach==6.2.0
blinker==1.4
Brlapi==0.8.3
celery==5.3.4
certifi==2025.8.3
cffi==2.0.0
chardet==4.0.0
charset-normalizer==3.4.3
chrome-gnome-shell==0.0.0
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
command-not-found==0.3
coverage==7.10.6
cryptography==46.0.1
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distro==1.7.0
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
exceptiongroup==1.3.0
fastapi==0.104.1
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.5
flake8==6.1.0
frozenlist==1.7.0
ghp-import==2.1.0
greenlet==3.2.4
gyp==0.1
h11==0.16.0
hidpidaemon==18.4.6
httpcore==1.0.9
httplib2==0.20.2
httptools==0.6.4
httpx==0.25.2
idna==3.10
importlib-metadata==4.6.4
iniconfig==2.1.0
isort==5.12.0
jeepney==0.7.1
Jinja2==3.1.6
kernelstub==3.1.4
keyring==23.5.0
kombu==5.5.4
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.3.10
Markdown==3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mccabe==0.7.0
mdurl==0.1.2
mergedeep==1.3.4
mkdocs==1.5.3
mkdocs-material==9.4.8
mkdocs-material-extensions==1.3.1
more-itertools==8.10.0
multidict==6.6.4
mypy_extensions==1.1.0
netaddr==0.8.0
netifaces==0.11.0
oauthlib==3.2.0
packaging==25.0
paginate==0.5.7
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.4.0
pluggy==1.6.0
podman-compose==1.5.0
pop-transition==1.1.2
prompt_toolkit==3.0.52
propcache==0.3.2
protobuf==3.12.4
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycairo==1.20.1
pycodestyle==2.11.1
pycparser==2.23
pycups==2.0.1
pydantic==2.11.9
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydbus==0.6.0
pyflakes==3.1.0
Pygments==2.19.2
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
pymdown-extensions==10.16.1
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.0
python-gnupg==0.4.8
python-http-client==3.3.7
python-jose==3.3.0
python-multipart==0.0.6
python-xlib==0.29
pytz==2022.1
pyxdg==0.27
PyYAML==6.0.2
pyyaml_env_tag==1.1
redis==5.0.1
regex==2025.9.1
repolib==2.2.1
repoman==1.4.0
requests==2.32.5
rich==14.1.0
rich-toolkit==0.14.9
rignore==0.6.4
rsa==4.9.1
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
sendgrid==6.12.4
sentry-sdk==2.34.1
sessioninstaller==0.0.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.27.0
structlog==23.2.0
systemd-python==234
tomli==2.2.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
ubuntu-drivers-common==0.0.0
ufw==0.36.1
urllib3==2.5.0
uvicorn==0.24.0
uvloop==0.21.0
vine==5.1.0
wadllib==1.3.6
watchdog==6.0.0
watchfiles==1.1.0
watchgod==0.8.2
wcwidth==0.2.13
webencodings==0.5.1
websockets==15.0.1
Werkzeug==3.1.3
wsproto==1.0.0
xdg==5
xkit==0.0.0
yarl==1.20.1
zipp==1.0.0
africastalking==1.2.5
